[Feat]: UserInfo Model 구성
